requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
PyJWT